src/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
decorator==4.4.2
Django==5.1.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-modeltranslation==0.19.9
django-polymorphic==3.1.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-yasg==1.21.7
email_validator==2.2.0
ffmpeg-python==0.2.0
future==1.0.0
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
inflection==0.5.1
kombu==5.4.0
moviepy==1.0.3
numpy==2.1.1
packaging==24.1
pillow==10.4.0
proglog==0.1.10
prompt_toolkit==3.0.47
psycopg==3.2.1
psycopg2==2.9.9
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
